bug fixes per review feedback and additional tests
